global memory file map handling separated into new unit ecal_memfile_map
