+++ ecal/core/CMakeLists.txt
@@ -57,6 +57,7 @@ set(ecal_io_cpp_src
 if(NOT ECAL_LAYER_ICEORYX)
   set(ecal_io_mem_cpp_src
       src/io/ecal_memfile.cpp
+      src/io/ecal_memfile_map.cpp
       src/io/ecal_memfile_pool.cpp
       src/io/ecal_memfile_sync.cpp
   )
@@ -172,6 +173,7 @@ if(NOT ECAL_LAYER_ICEORYX)
   set(ecal_io_mem_header_src
       src/io/ecal_memfile.h
       src/io/ecal_memfile_mtx.h
+      src/io/ecal_memfile_map.h
       src/io/ecal_memfile_pool.h
       src/io/ecal_memfile_sync.h
   )
