+++ ecal/core/src/io/ecal_memfile.cpp
@@ -21,31 +21,19 @@
  * @brief  memory file interface
 **/
 
-#include <ecal/ecal_os.h>
-
 #include "ecal_def.h"
+#include "ecal_global_accessors.h"
 #include "ecal_memfile.h"
+#include "ecal_memfile_map.h"
 
-#include <stdio.h>
-#include <string.h>
-#include <iostream>
 #include <assert.h>
 
 namespace eCAL
 {
   /////////////////////////////////////////////////////////////////////////////////
-  // Memory file handling util functions
+  // External memory file handling functions (ecal_memfile_win32|linux.cpp)
   /////////////////////////////////////////////////////////////////////////////////
-
-  bool CreateMemFile(const std::string& name_, const bool create_, const size_t len_, SMemFileInfo& mem_file_info_);
-  bool AllocMemFile(const std::string& name_, const bool create_, SMemFileInfo& mem_file_info_);
-  bool MapMemFile(const bool create_, SMemFileInfo& mem_file_info_);
   bool CheckMemFile(const size_t len_, const bool create_, SMemFileInfo& mem_file_info_);
-  bool UnMapMemFile(SMemFileInfo& mem_file_info_);
-  bool DeAllocMemFile(SMemFileInfo& mem_file_info_);
-  bool DestroyMemFile(const std::string& name_, const bool remove_);
-  bool RemoveMemFile(const SMemFileInfo& mem_file_info_);
-
 
   /////////////////////////////////////////////////////////////////////////////////
   // Memory file handling class
@@ -329,10 +317,7 @@ namespace eCAL
       CheckMemFile(len, false, m_memfile_info);
 
       // update memory file map
-      {
-        std::lock_guard<std::mutex> lock(g_memfile_map()->sync);
-        g_memfile_map()->map[m_name] = m_memfile_info;
-      }
+      UpdateMemFile(m_name, m_memfile_info);
 
       // check size again and give up if it is still to small
       if (len > m_memfile_info.size)
@@ -344,127 +329,5 @@ namespace eCAL
     }
 
     return(true);
-  }
-
-  /////////////////////////////////////////////////////////////////////////////////
-  // Memory file global map handling
-  /////////////////////////////////////////////////////////////////////////////////
-
-  void CleanupMemoryFileMap()
-  {
-    if (!g_memfile_map()) return;
-
-    // lock memory map access
-    std::lock_guard<std::mutex> lock(g_memfile_map()->sync);
-
-    // erase memory files from memory map
-    auto& memfile_map = g_memfile_map()->map;
-    for (MemFileMapT::iterator iter = memfile_map.begin(); iter != memfile_map.end(); ++iter)
-    {
-      auto& memfile_info = iter->second;
-
-      // unmap memory file
-      UnMapMemFile(memfile_info);
-
-      // remove memory file from system
-      if (memfile_info.remove) RemoveMemFile(memfile_info);
-
-      // deallocate memory file
-      DeAllocMemFile(memfile_info);
-    }
-
-    // clear map
-    memfile_map.clear();
-  }
-
-  /////////////////////////////////////////////////////////////////////////////////
-  // Memory file handling help functions
-  /////////////////////////////////////////////////////////////////////////////////
-
-  bool CreateMemFile(const std::string& name_, const bool create_, const size_t len_, SMemFileInfo& mem_file_info_)
-  {
-    if (!g_memfile_map()) return(false);
-
-    // we need a length != 0
-    assert(len_ > 0);
-
-    // lock memory map access
-    std::lock_guard<std::mutex> lock(g_memfile_map()->sync);
-
-    // check for existing memory file
-    MemFileMapT::iterator iter = g_memfile_map()->map.find(name_);
-    if (iter == g_memfile_map()->map.end())
-    {
-      // create memory file
-      if (!AllocMemFile(name_, create_, mem_file_info_))
-      {
-#ifndef NDEBUG
-        printf("Could create memory file: %s.\n\n", name_.c_str());
-#endif
-        return(false);
-      }
-
-      // check memory file size
-      CheckMemFile(len_, create_, mem_file_info_);
-
-      // and add to memory file map
-      mem_file_info_.refcnt++;
-      g_memfile_map()->map[name_] = mem_file_info_;
-    }
-    else
-    {
-      // increase reference counter
-      iter->second.refcnt++;
-
-      // check memory file size
-      CheckMemFile(len_, false, iter->second);
-
-      // copy info from memory file map
-      mem_file_info_ = iter->second;
-    }
-
-    // return success
-    return(true);
-  }
-
-  bool DestroyMemFile(const std::string& name_, const bool remove_)
-  {
-    if (!g_memfile_map()) return(false);
-
-    // lock memory map access
-    std::lock_guard<std::mutex> lock(g_memfile_map()->sync);
-
-    // erase memory file from memory map
-    auto& memfile_map = g_memfile_map()->map;
-    MemFileMapT::iterator iter = memfile_map.find(name_);
-    if (iter != memfile_map.end())
-    {
-      auto& memfile_info = iter->second;
-
-      // decrease reference counter
-      memfile_info.refcnt--;
-      // mark for remove
-      memfile_info.remove |= remove_;
-      if (memfile_info.refcnt < 1)
-      {
-        bool remove_from_system = memfile_info.remove;
-
-        // unmap memory file
-        UnMapMemFile(memfile_info);
-
-        // remove memory file from system
-        if (remove_from_system) RemoveMemFile(memfile_info);
-
-        // dealloc memory file
-        DeAllocMemFile(memfile_info);
-
-        memfile_map.erase(iter);
-
-        // we removed the file
-        return(true);
-      }
-    }
-
-    return(false);
   }
 }
