+++ ecal/core/src/io/ecal_memfile_map.cpp
@@ -1,0 +1,171 @@
+/* ========================= eCAL LICENSE =================================
+ *
+ * Copyright (C) 2016 - 2019 Continental Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * ========================= eCAL LICENSE =================================
+*/
+
+/**
+ * @brief  eCAL memory file map handling
+**/
+
+#include <ecal/ecal_os.h>
+
+#include "ecal_def.h"
+#include "ecal_global_accessors.h"
+#include "ecal_memfile_map.h"
+
+#include <assert.h>
+
+namespace eCAL
+{
+  /////////////////////////////////////////////////////////////////////////////////
+  // External memory file handling functions (ecal_memfile_win32|linux.cpp)
+  /////////////////////////////////////////////////////////////////////////////////
+  bool AllocMemFile(const std::string& name_, const bool create_, SMemFileInfo& mem_file_info_);
+  bool CheckMemFile(const size_t len_, const bool create_, SMemFileInfo& mem_file_info_);
+  bool UnMapMemFile(SMemFileInfo& mem_file_info_);
+  bool DeAllocMemFile(SMemFileInfo& mem_file_info_);
+  bool RemoveMemFile(const SMemFileInfo& mem_file_info_);
+    
+#if 0 // NOT USED ?!
+  void CleanupMemoryFileMap()
+  {
+    if (!g_memfile_map()) return;
+
+    // lock memory map access
+    std::lock_guard<std::mutex> lock(g_memfile_map()->sync);
+
+    // erase memory files from memory map
+    auto& memfile_map = g_memfile_map()->map;
+    for (MemFileMapT::iterator iter = memfile_map.begin(); iter != memfile_map.end(); ++iter)
+    {
+      auto& memfile_info = iter->second;
+
+      // unmap memory file
+      UnMapMemFile(memfile_info);
+
+      // remove memory file from system
+      if (memfile_info.remove) RemoveMemFile(memfile_info);
+
+      // deallocate memory file
+      DeAllocMemFile(memfile_info);
+    }
+
+    // clear map
+    memfile_map.clear();
+  }
+#endif
+
+  bool CreateMemFile(const std::string& name_, const bool create_, const size_t len_, SMemFileInfo& mem_file_info_)
+  {
+    if (!g_memfile_map()) return(false);
+
+    // we need a length != 0
+    assert(len_ > 0);
+
+    // lock memory map access
+    std::lock_guard<std::mutex> lock(g_memfile_map()->sync);
+
+    // check for existing memory file
+    MemFileMapT::iterator iter = g_memfile_map()->map.find(name_);
+    if (iter == g_memfile_map()->map.end())
+    {
+      // create memory file
+      if (!AllocMemFile(name_, create_, mem_file_info_))
+      {
+#ifndef NDEBUG
+        printf("Could create memory file: %s.\n\n", name_.c_str());
+#endif
+        return(false);
+      }
+
+      // check memory file size
+      CheckMemFile(len_, create_, mem_file_info_);
+
+      // and add to memory file map
+      mem_file_info_.refcnt++;
+      g_memfile_map()->map[name_] = mem_file_info_;
+    }
+    else
+    {
+      // increase reference counter
+      iter->second.refcnt++;
+
+      // check memory file size
+      CheckMemFile(len_, false, iter->second);
+
+      // copy info from memory file map
+      mem_file_info_ = iter->second;
+    }
+
+    // return success
+    return(true);
+  }
+
+  bool DestroyMemFile(const std::string& name_, const bool remove_)
+  {
+    if (!g_memfile_map()) return(false);
+
+    // lock memory map access
+    std::lock_guard<std::mutex> lock(g_memfile_map()->sync);
+
+    // erase memory file from memory map
+    auto& memfile_map = g_memfile_map()->map;
+    MemFileMapT::iterator iter = memfile_map.find(name_);
+    if (iter != memfile_map.end())
+    {
+      auto& memfile_info = iter->second;
+
+      // decrease reference counter
+      memfile_info.refcnt--;
+      // mark for remove
+      memfile_info.remove |= remove_;
+      if (memfile_info.refcnt < 1)
+      {
+        bool remove_from_system = memfile_info.remove;
+
+        // unmap memory file
+        UnMapMemFile(memfile_info);
+
+        // remove memory file from system
+        if (remove_from_system) RemoveMemFile(memfile_info);
+
+        // dealloc memory file
+        DeAllocMemFile(memfile_info);
+
+        memfile_map.erase(iter);
+
+        // we removed the file
+        return(true);
+      }
+    }
+
+    return(false);
+  }
+
+  bool UpdateMemFile(const std::string& name_, SMemFileInfo& mem_file_info_)
+  {
+    if (!g_memfile_map()) return(false);
+
+    // lock memory map access
+    std::lock_guard<std::mutex> lock(g_memfile_map()->sync);
+
+    // update/set info
+    g_memfile_map()->map[name_] = mem_file_info_;
+
+    return(true);
+  }
+}
